コードが肥大化してきたのでリファクタリング
